requirements.txt
notion-client>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
google-genai
